node:util: move deprecate to internal file so its faster to import
