+++ src/js/node/util.ts
@@ -5,6 +5,7 @@ const utl = require("internal/util/inspect");
 const { promisify } = require("internal/promisify");
 const { validateString, validateOneOf } = require("internal/validators");
 const { MIMEType, MIMEParams } = require("internal/util/mime");
+const { deprecate } = require("internal/util/deprecate");
 
 const internalErrorName = $newZigFunction("node_util_binding.zig", "internalErrorName", 1);
 const parseEnv = $newZigFunction("node_util_binding.zig", "parseEnv", 1);
@@ -30,46 +31,6 @@ const inspect = utl.inspect;
 const formatWithOptions = utl.formatWithOptions;
 const format = utl.format;
 const stripVTControlCharacters = utl.stripVTControlCharacters;
-
-const codesWarned = new Set();
-
-function getDeprecationWarningEmitter(code, msg, deprecated, shouldEmitWarning = () => true) {
-  let warned = false;
-  return function () {
-    if (!warned && shouldEmitWarning()) {
-      warned = true;
-      if (code !== undefined) {
-        if (!codesWarned.has(code)) {
-          process.emitWarning(msg, "DeprecationWarning", code, deprecated);
-          codesWarned.add(code);
-        }
-      } else {
-        process.emitWarning(msg, "DeprecationWarning", deprecated);
-      }
-    }
-  };
-}
-
-function deprecate(fn, msg, code) {
-  // Lazy-load to avoid a circular dependency.
-  if (code !== undefined) validateString(code, "code");
-
-  const emitDeprecationWarning = getDeprecationWarningEmitter(code, msg, deprecated);
-
-  function deprecated(...args) {
-    if (!process.noDeprecation) {
-      emitDeprecationWarning();
-    }
-    if (new.target) {
-      return Reflect.construct(fn, args, new.target);
-    }
-    return fn.$apply(this, args);
-  }
-
-  // The wrapper will keep the same prototype as fn to maintain prototype chain
-  Object.setPrototypeOf(deprecated, fn);
-  return deprecated;
-}
 
 var debugs = {};
 var debugEnvRegex = /^$/;
